requirements.txt
pdfminer.six
pandas
boto3
requests
